refactor: move channel fact authority to perimeter
