+++ packages/channels/src/router/channel-grant.ts
@@ -1,0 +1,100 @@
+import type { Actor } from "@openomni/protocol";
+import { ChannelGrantStore } from "@openomni/ledger";
+
+export interface ChannelGrantMatchInput {
+  readonly surface: string;
+  readonly workspace?: string;
+  readonly channel?: string;
+}
+
+export interface ChannelGrantResolution {
+  readonly grant: Actor.ChannelGrant;
+  readonly inboundTreatment: Actor.InboundTreatment;
+}
+
+function defaultTreatment(kind: Actor.ChannelGrantKind): Actor.InboundTreatment {
+  if (kind === "trusted_channel") return "full_access";
+  if (kind === "broadcast_channel") return "evidence_only";
+  return "drop";
+}
+
+function matches(grant: Actor.ChannelGrant, input: ChannelGrantMatchInput): boolean {
+  if (grant.surface !== input.surface) return false;
+  if (grant.workspace !== undefined && grant.workspace !== input.workspace) return false;
+  if (grant.channel !== undefined && grant.channel !== input.channel) return false;
+  return true;
+}
+
+function specificity(grant: Actor.ChannelGrant): number {
+  return (grant.workspace === undefined ? 0 : 1) + (grant.channel === undefined ? 0 : 1);
+}
+
+const treatmentRestriction: Record<Actor.InboundTreatment, number> = {
+  drop: 0,
+  evidence_only: 1,
+  full_access: 2,
+};
+
+const defaultTierRestriction: Record<Actor.TrustTier, number> = {
+  assigned_worker: 1,
+  observer: 2,
+  collaborator: 3,
+  manager: 4,
+  co_owner: 5,
+  owner: 6,
+};
+
+const kindRestriction: Record<Actor.ChannelGrantKind, number> = {
+  blocked_channel: 0,
+  broadcast_channel: 1,
+  trusted_channel: 2,
+};
+
+function effectiveTreatment(grant: Actor.ChannelGrant): Actor.InboundTreatment {
+  const treatment = grant.inboundTreatment ?? defaultTreatment(grant.kind);
+  if (grant.kind === "blocked_channel" || treatment === "drop") return "drop";
+  if (grant.kind === "broadcast_channel") return "evidence_only";
+  return treatment;
+}
+
+function compareStableText(a: string, b: string): number {
+  if (a < b) return -1;
+  if (a > b) return 1;
+  return 0;
+}
+
+/**
+ * Chooses the perimeter's effective channel grant. Specificity wins first;
+ * equal-score grants use a fail-closed, backend-independent total order.
+ */
+function compareResolutionOrder(a: Actor.ChannelGrant, b: Actor.ChannelGrant): number {
+  const specificityOrder = specificity(b) - specificity(a);
+  if (specificityOrder !== 0) return specificityOrder;
+
+  const treatmentOrder =
+    treatmentRestriction[effectiveTreatment(a)] - treatmentRestriction[effectiveTreatment(b)];
+  if (treatmentOrder !== 0) return treatmentOrder;
+
+  const defaultTierOrder =
+    (a.defaultTier === undefined ? 0 : defaultTierRestriction[a.defaultTier]) -
+    (b.defaultTier === undefined ? 0 : defaultTierRestriction[b.defaultTier]);
+  if (defaultTierOrder !== 0) return defaultTierOrder;
+
+  const kindOrder = kindRestriction[a.kind] - kindRestriction[b.kind];
+  if (kindOrder !== 0) return kindOrder;
+
+  return compareStableText(a.id, b.id);
+}
+
+export function resolveChannelGrant(
+  input: ChannelGrantMatchInput,
+): ChannelGrantResolution | undefined {
+  const grants = ChannelGrantStore.list().filter((grant) => matches(grant, input));
+  grants.sort(compareResolutionOrder);
+  const grant = grants[0];
+  if (grant === undefined) return undefined;
+  return {
+    grant,
+    inboundTreatment: effectiveTreatment(grant),
+  };
+}
